refactor: split up global tables into structs
